moved summary icon code
