+++ src/components/SummaryTable.vue
@@ -2,7 +2,7 @@
   <q-overlay
     class="bg-secondary"
     :z-index="5000"
-    :summaryOverlay="summaryOverlay">
+  >
     <template #body>
       <div class="justify-evenly summaryContainer">
         <!-- {{ content }} -->
@@ -18,7 +18,6 @@
 // import { evaluate } from '../ts/main/quiz/evaluate/evaluate.support'
 import SummaryExit from './SummaryExit.vue';
 
-defineProps({ summaryOverlay: Boolean })
 const emit = defineEmits(['closeSummary'])
 function closeSummary() {
   emit('closeSummary')
